// src/app/shared/models/Usuario/Empleado/RegistroEmpleado.ts
export class RegistroEmpleado{
    username: string;
    password: string;
    direccion: string;
    estado?: boolean;
    ciudad_id: number;
    creador_id: number;
    nombre_completo: string;
    ci: string;
    fecha_nacimiento: string;
    nacionalidad: string;
    estado_civil: string;
    interno: boolean;
}